Issue SKER4939:  (Fix memory leak in sesat commons ReferenceMap)
Changed the ReferenceMap implementation to use only one thread. This thread is created in a static way together with the ReferenceQueue that keeps track of the Weak/Soft references that has had it's object garbage collected. The thread will only know about the static queue, and not as before the queue belonging to a specific ReferenceMap object. The hard reference from the thread to the ReferenceMap's queue is therefor removed, and the memory leak is gone. This will also lead to improved performance since we have less switching between threads.